+++ Sources/Sentry/SentryEnvelope.m
@@ -93,6 +93,8 @@ NS_ASSUME_NONNULL_BEGIN
     // default. In any case in the envelope type it should be event. Except for transactions
     NSString *envelopeType = [event.type isEqualToString:SentryEnvelopeItemTypeTransaction]
         ? SentryEnvelopeItemTypeTransaction
+        : [event.type isEqualToString:SentryEnvelopeItemTypeFeedback]
+        ? SentryEnvelopeItemTypeFeedback
         : SentryEnvelopeItemTypeEvent;
 
     return [self initWithHeader:[[SentryEnvelopeItemHeader alloc] initWithType:envelopeType
@@ -159,7 +161,17 @@ NS_ASSUME_NONNULL_BEGIN
             return nil;
         }
 
+#if DEBUG || SENTRY_TEST || SENTRY_TEST_CI
+        if ([NSProcessInfo.processInfo.arguments
+                containsObject:@"--io.sentry.base64-attachment-data"]) {
+            data = [[attachment.data base64EncodedStringWithOptions:0]
+                dataUsingEncoding:NSUTF8StringEncoding];
+        } else {
+            data = attachment.data;
+        }
+#else
         data = attachment.data;
+#endif // DEBUG || SENTRY_TEST || SENTRY_TEST_CI
     } else if (nil != attachment.path) {
 
         NSError *error = nil;
@@ -184,7 +196,17 @@ NS_ASSUME_NONNULL_BEGIN
             return nil;
         }
 
+#if DEBUG || SENTRY_TEST || SENTRY_TEST_CI
+        if ([NSProcessInfo.processInfo.arguments
+                containsObject:@"--io.sentry.base64-attachment-data"]) {
+            data = [[[[NSFileManager defaultManager] contentsAtPath:attachment.path]
+                base64EncodedStringWithOptions:0] dataUsingEncoding:NSUTF8StringEncoding];
+        } else {
+            data = [[NSFileManager defaultManager] contentsAtPath:attachment.path];
+        }
+#else
         data = [[NSFileManager defaultManager] contentsAtPath:attachment.path];
+#endif // DEBUG || SENTRY_TEST || SENTRY_TEST_CI
     }
 
     if (data == nil) {