+++ Sources/Swift/Integrations/SessionReplay/RRWeb/SentryRRWebOptionsEvent.swift
@@ -3,17 +3,27 @@ import Foundation
 
 @objc class SentryRRWebOptionsEvent: SentryRRWebCustomEvent {
     
+    init(timestamp: Date, customOptions: [String: Any]) {
+        super.init(timestamp: timestamp, tag: "options", payload: customOptions)
+    }
+    
     init(timestamp: Date, options: SentryReplayOptions) {
-        super.init(timestamp: timestamp, tag: "options", payload:
-                    [
-                        "sessionSampleRate": options.sessionSampleRate,
-                        "errorSampleRate": options.onErrorSampleRate,
-                        "maskAllText": options.maskAllText,
-                        "maskAllImages": options.maskAllImages,
-                        "quality": String(describing: options.quality),
-                        "maskedViewClasses": options.maskedViewClasses.map(String.init(describing: )).joined(separator: ", "),
-                        "unmaskedViewClasses": options.unmaskedViewClasses.map(String.init(describing: )).joined(separator: ", ")
-                    ]
-        )
+        var payload: [String: Any] = [
+            "sessionSampleRate": options.sessionSampleRate,
+            "errorSampleRate": options.onErrorSampleRate,
+            "maskAllText": options.maskAllText,
+            "maskAllImages": options.maskAllImages,
+            "quality": String(describing: options.quality)
+        ]
+        
+        if !options.maskedViewClasses.isEmpty {
+            payload["maskedViewClasses"] = options.maskedViewClasses.map(String.init(describing:)).joined(separator: ", ")
+        }
+        
+        if !options.unmaskedViewClasses.isEmpty {
+            payload["unmaskedViewClasses"] = options.unmaskedViewClasses.map(String.init(describing:)).joined(separator: ", ")
+        }
+        
+        super.init(timestamp: timestamp, tag: "options", payload: payload)
     }
 }