+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackForm.swift
@@ -8,8 +8,7 @@ import UIKit
 
 @available(iOS 13.0, *)
 protocol SentryUserFeedbackFormDelegate: NSObjectProtocol {
-    func cancelled()
-    func confirmed()
+    func finished(with feedback: SentryFeedback?)
 }
 
 @available(iOS 13.0, *)
@@ -26,6 +25,7 @@ class SentryUserFeedbackForm: UIViewController {
         updateLayout()
     }
     
+    //swiftlint:disable function_body_length
     init(config: SentryUserFeedbackConfiguration, delegate: any SentryUserFeedbackFormDelegate) {
         self.config = config
         self.delegate = delegate
@@ -83,12 +83,13 @@ class SentryUserFeedbackForm: UIViewController {
             $0.setTitleColor(config.theme.buttonForeground, for: .normal)
         }
     }
+    //swiftlint:enable function_body_length
     
     // MARK: Actions
     
     func addScreenshotButtonTapped() {
         // the iOS photo picker UI doesn't play nicely with XCUITest, so we'll just mock the selection here
-#if TEST || TESTCI
+#if SENTRY_TEST || SENTRY_TEST_CI
         //swiftlint:disable force_try force_unwrapping
         let url = Bundle.main.url(forResource: "Tongariro", withExtension: "jpg")!
         let image = try! UIImage(data: Data(contentsOf: url))!
@@ -101,7 +102,7 @@ class SentryUserFeedbackForm: UIViewController {
         imagePickerController.sourceType = .photoLibrary
         imagePickerController.allowsEditing = true
         present(imagePickerController, animated: config.animations)
-#endif // TEST || TESTCI
+#endif // SENTRY_TEST || SENTRY_TEST_CI
     }
     
     func removeScreenshotButtonTapped() {
@@ -129,15 +130,25 @@ class SentryUserFeedbackForm: UIViewController {
             let list = missing.count == 1 ? missing[0] : missing[0 ..< missing.count - 1].joined(separator: ", ") + " and " + missing[missing.count - 1]
             let alert = UIAlertController(title: "Error", message: "You must provide all required information. Please check the following field\(missing.count > 1 ? "s" : ""): \(list).", preferredStyle: .alert)
             alert.addAction(UIAlertAction(title: "OK", style: .default))
-            present(alert, animated: config.animations)
+            present(alert, animated: config.animations) {
+                if let block = self.config.onSubmitError {
+                    // we use NSError here instead of Swift.Error because NSError automatically bridges to Swift.Error, but the same is not true in the other direction if you want to include a userInfo dictionary. Using Swift.Error would require additional implementation for this to work with ObjC consumers.
+                    block(NSError(domain: "io.sentry.error", code: 1, userInfo: ["missing_fields": missing, NSLocalizedDescriptionKey: "The user did not complete the feedback form."]))
+                }
+            }
             return
         }
-        
-        delegate?.confirmed()
+
+        let feedback = SentryFeedback(message: messageTextView.text, name: fullNameTextField.text, email: emailTextField.text, screenshot: screenshotImageView.image?.pngData())
+        SentryLog.log(message: "Sending user feedback", andLevel: .debug)
+        if let block = config.onSubmitSuccess {
+            block(feedback.dataDictionary())
+        }
+        delegate?.finished(with: feedback)
     }
     
     func cancelButtonTapped() {
-        delegate?.cancelled()
+        delegate?.finished(with: nil)
     }
     
     // MARK: Layout