+++ Sources/Swift/Integrations/UserFeedback/SentryUserFeedbackWidget.swift
@@ -7,6 +7,10 @@ import UIKit
 
 var displayingForm = false
 
+protocol SentryUserFeedbackWidgetDelegate: NSObjectProtocol {
+    func capture(feedback: SentryFeedback)
+}
+
 @available(iOS 13.0, *)
 struct SentryUserFeedbackWidget {
     class Window: UIWindow {
@@ -17,13 +21,18 @@ struct SentryUserFeedbackWidget {
                 self.setWidget(visible: false)
                 let form = SentryUserFeedbackForm(config: self.config, delegate: self)
                 form.presentationController?.delegate = self
-                self.present(form, animated: self.config.animations)
+                self.present(form, animated: self.config.animations) {  
+                    self.config.onFormOpen?()
+                }
             })
             
             let config: SentryUserFeedbackConfiguration
             
-            init(config: SentryUserFeedbackConfiguration) {
+            weak var delegate: (any SentryUserFeedbackWidgetDelegate)?
+            
+            init(config: SentryUserFeedbackConfiguration, delegate: any SentryUserFeedbackWidgetDelegate) {
                 self.config = config
+                self.delegate = delegate
                 super.init(nibName: nil, bundle: nil)
                 view.addSubview(button)
                 
@@ -63,32 +72,32 @@ struct SentryUserFeedbackWidget {
             
             func closeForm() {
                 setWidget(visible: true)
-                dismiss(animated: config.animations)
+                dismiss(animated: config.animations, completion: {
+                    self.config.onFormClose?()
+                })
             }
             
             // MARK: SentryUserFeedbackFormDelegate
             
-            func cancelled() {
+            func finished(with feedback: SentryFeedback?) {
                 closeForm()
+                
+                if let feedback = feedback {
+                    delegate?.capture(feedback: feedback)
+                }
             }
-            
-//swiftlint:disable todo
-            func confirmed() {
-                // TODO: submit
-                closeForm()
-            }
-//swiftlint:enable todo
             
             // MARK: UIAdaptivePresentationControllerDelegate
             
             func presentationControllerDidDismiss(_ presentationController: UIPresentationController) {
                 setWidget(visible: true)
+                self.config.onFormClose?()
             }
         }
         
-        init(config: SentryUserFeedbackConfiguration) {
+        init(config: SentryUserFeedbackConfiguration, delegate: any SentryUserFeedbackWidgetDelegate) {
             super.init(frame: UIScreen.main.bounds)
-            rootViewController = RootViewController(config: config)
+            rootViewController = RootViewController(config: config, delegate: delegate)
             windowLevel = config.widgetConfig.windowLevel
         }